+++ docs/content/templates/content/index.mdx
@@ -5,7 +5,7 @@ description: The content template is used to deliver long-form informational con
 version: 1.0.0
 previewPath: /category/subcategory/content
 figmaTemplateNodeId: 8410%3A83580
-githubTemplatePath: /example-site/pages/category/subcategory/content.tsx
+githubTemplatePath: /docs/content/templates/content/Content.tsx
 storybookPath: /story/templates-content-page--basic
 ---
